1.8 - BETA7
Added smoother scaling
Added an option to keep rendering the overlay in chat

+++ src/main/java/ga/matthewtgm/simplehud/elements/impl/ElementCPS.java
@@ -25,7 +25,7 @@ public class ElementCPS extends Element {
     @Override
     public void onRendered(ElementPosition position) {
         this.setRenderedValue(this.getCPS());
-        this.height = 10 * this.getPosition().getScale();
+        this.height = (int) (10 * this.getPosition().getScale());
         super.onRendered(position);
     }
