+++ src/main/java/ga/matthewtgm/simplehud/elements/impl/ElementSimpleText.java
@@ -86,7 +86,7 @@ public class ElementSimpleText extends Element {
     @Override
     public void onRendered(ElementPosition position) {
         this.setRenderedValue(this.getText());
-        this.height = 10 * this.getPosition().getScale();
+        this.height = (int) (10 * this.getPosition().getScale());
         super.onRendered(position);
     }
